Workarounds for problems, static widget layout
First of all, here are the solutions found to the problems encountered
in the last commit:

* for "Content Not Allowed In Prolog" build errors

    just delete the auto-generated res/playstore and res/LICENSE dirs
    and rebuild

* for instant crash in Android Studio launch

    point JAVA_HOME, STUDIO_JDK and JDK_HOME to the /usr/lib/jre... the
    system's openjdk already installed (e.g. 1.8.0_121 not the Android
    Studio bundled one 1.8.0_76)

    launch Studio with another project (cancel default project loading
    in startup), close this one (the problematic) and reopen it.

    If not fixed with above, try to change the project's Sdk location to
    the new one already mentioned not the bundled one in File > Project
    Settings category

Finally, some more work is done regarding the widget's look. The Layout
to be used must be A Linear one, dynamically populated with the csv
contents. This can also be made with an Adapter (and why not a GridView
as stated in the tutorial)